+++ packages/stash/src/actions/extend.ts
@@ -1,0 +1,15 @@
+import { Stash } from "../common";
+
+export type ExtendArgs<stash extends Stash, actions> = {
+  stash: stash;
+  actions: actions;
+};
+
+export type ExtendResult<stash extends Stash, actions> = Omit<stash, keyof actions> & actions;
+
+export function extend<stash extends Stash, actions>({
+  stash,
+  actions,
+}: ExtendArgs<stash, actions>): ExtendResult<stash, actions> {
+  return { ...stash, ...actions };
+}