+++ test/mock-game-contracts/CHANGELOG.md
@@ -1,5 +1,7 @@
 # mock-game-contracts
 
+## 2.2.0
+
 ## 2.1.1
 
 ## 2.1.0